+++ transformer/src/getMethods.ts
@@ -9,6 +9,7 @@ import { getSignatureParameters } from "./getSignatureParameters";
 import { getTypeCall }            from "./getTypeCall";
 import {
 	getAccessModifier,
+	getDeclaration,
 	getFunctionLikeSignature,
 	getUnknownTypeCall,
 }                                 from "./helpers";
@@ -23,26 +24,16 @@ export function getMethodGenerics(symbol: ts.Symbol, context: Context): Array<Ge
 	});
 }
 
-/**
- * Extracted to its own method, hopefully we should be able to use
- * this for other method descriptions, than just a class method
- *
- * @param {ts.Symbol} symbol
- * @param {ts.Declaration} declaration
- * @param {Context} context
- * @returns {MethodDescriptionSource}
- */
-export function getMethodDescription(symbol: ts.Symbol, declaration: ts.Declaration, context: Context): MethodDescriptionSource
+function getMethodDescriptionFromSignature(symbol: ts.Symbol, methodSignature: ts.Signature, context: Context)
 {
-	const methodSignature = getFunctionLikeSignature(symbol, declaration, context.typeChecker);
-	const returnType = methodSignature?.getReturnType();
+	const returnType = methodSignature.getReturnType();
 
 	return {
 		n: symbol.escapedName.toString(),
-		params: methodSignature && getSignatureParameters(methodSignature, context),
-		rt: returnType && getTypeCall(returnType, returnType.symbol, context) || getUnknownTypeCall(context),
+		params: getSignatureParameters(methodSignature, context),
+		rt: getTypeCall(returnType, returnType.symbol, context) || getUnknownTypeCall(context),
 		d: getDecorators(symbol, context),
-		tp: getMethodGenerics(symbol, context),
+		tp: methodSignature.getTypeParameters()?.map(typeParameter => getTypeCall(typeParameter, typeParameter.symbol, context)) ?? [], //getMethodGenerics(symbol, context),
 		o: (symbol.flags & ts.SymbolFlags.Optional) === ts.SymbolFlags.Optional,
 		am: getAccessModifier(symbol.valueDeclaration?.modifiers)
 	};
@@ -66,9 +57,23 @@ export function getMethods(symbol: ts.Symbol | undefined, type: ts.Type, context
 	const methods = members
 		.filter(m => (m.flags & ts.SymbolFlags.Method) === ts.SymbolFlags.Method || (m.flags & ts.SymbolFlags.Function) === ts.SymbolFlags.Function)
 		.flatMap(
-			(memberSymbol: ts.Symbol) => memberSymbol.getDeclarations()?.map(
-				declaration => getMethodDescription(memberSymbol, declaration, context)
-			) ?? []
+			(memberSymbol: ts.Symbol) => {
+				const declaration = getDeclaration(memberSymbol);
+
+				if (!declaration)
+				{
+					return [];
+				}
+				
+				let type = context.typeChecker.getTypeOfSymbolAtLocation(memberSymbol, declaration);
+				
+				if (type.isUnion()) 
+				{
+					type = (type.types[0].flags === ts.TypeFlags.Undefined ? type.types[1] : type.types[0]) || type;
+				}
+				
+				return type.getCallSignatures().map(signature => getMethodDescriptionFromSignature(memberSymbol, signature, context));
+			}
 		);
 
 	return methods.length ? methods : undefined;