+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [//]: # (### Added)
 [//]: # (### Changed)
 
+## [0.12.0] - 2022-09-02 - transformer
+### Added
+- Method and function overloads,
+- `Type.toString()` - returns string with kind, name and fullName of the type.
+### Changed
+- Type.function replaced by Type.getSignatures();
+
 ## [0.11.0] - 2022-08-14 - transformer
 ### Added
 - Type.isSymbol(): boolean;